==> notebooks/03_text_preprocessing.ipynb.ipynb <==
import pandas as pd
import re
import nltk


from nltk.corpus import stopwords
from nltk.stem import WordNetLemmatizer


nltk.download('stopwords')
nltk.download('wordnet')


data_path = "../data/raw/Combined Data.csv"
df = pd.read_csv(data_path)
df.head()


df = df.dropna(subset=['statement'])
df.isnull().sum()


stop_words = set(stopwords.words('english'))
lemmatizer = WordNetLemmatizer()


def clean_text(text):
    # convert to lowercase
    text = text.lower()
    
    # remove urls
    text = re.sub(r'http\S+|www\S+', '', text)
    
    # remove non-alphabetic characters
    text = re.sub(r'[^a-z\s]', '', text)
    
    # tokenize
    words = text.split()
    
    # remove stopwords and lemmatize
    words = [lemmatizer.lemmatize(word) for word in words if word not in stop_words]
    
    return " ".join(words)


df['cleaned_statement'] = df['statement'].apply(clean_text)


df[['statement', 'cleaned_statement']].head()


df = df[df['cleaned_statement'].str.strip() != ""]


output_path = "../data/processed/cleaned_mental_health.csv"
df.to_csv(output_path, index=False)
