==> notebooks/04_emotion_detection.ipynb.ipynb <==
#pip install transformers torch (in cmd)

import pandas as pd
from transformers import pipeline


data_path = "../data/processed/cleaned_mental_health.csv"
df = pd.read_csv(data_path)
df.head()


df['status'].value_counts()


MAX_SAMPLES_PER_CLASS = 3000

def capped_sampling(group):
    if len(group) > MAX_SAMPLES_PER_CLASS:
        return group.sample(n=MAX_SAMPLES_PER_CLASS, random_state=42)
    else:
        return group  # take all if less than cap

df_capped = (
    df.groupby("status", group_keys=False)
      .apply(capped_sampling)
)

df_capped.shape


df_capped['status'].value_counts()


emotion_classifier = pipeline(
    "text-classification",
    model="j-hartmann/emotion-english-distilroberta-base",
    top_k=1,
    truncation=True,
    max_length=512
)


sample_text = df_capped['cleaned_statement'].iloc[0]
emotion_classifier(sample_text)


texts = df_capped['cleaned_statement'].tolist()

emotion_outputs = emotion_classifier(
    texts,
    batch_size=8,   # safe for CPU laptops
    truncation=True
)

df_capped['emotion'] = [result[0]['label'] for result in emotion_outputs]


df_capped[['cleaned_statement', 'emotion']].head()


df_capped['emotion'].value_counts()


output_path = "../data/processed/emotion_mental_health.csv"
df_capped.to_csv(output_path, index=False)


df_capped.shape


df_capped['status'].value_counts()
