==> notebooks/05_feature_engineering.ipynb.ipynb <==
import pandas as pd
import pickle
from sklearn.feature_extraction.text import TfidfVectorizer
from sklearn.preprocessing import LabelEncoder
from scipy.sparse import hstack


data_path = "../data/processed/emotion_mental_health.csv"
df = pd.read_csv(data_path)


df.head()


df.shape


df['status'].value_counts()


tfidf = TfidfVectorizer(
    max_features=5000,
    ngram_range=(1, 2)
)

X_text = tfidf.fit_transform(df['cleaned_statement'])
X_text.shape


emotion_encoder = LabelEncoder()
emotion_encoded = emotion_encoder.fit_transform(df['emotion'])
emotion_encoded = emotion_encoded.reshape(-1, 1)


X_features = hstack([X_text, emotion_encoded])
X_features.shape


label_encoder = LabelEncoder()
y = label_encoder.fit_transform(df['status'])


with open("../models/tfidf_vectorizer.pkl", "wb") as f:
    pickle.dump(tfidf, f)

with open("../models/emotion_encoder.pkl", "wb") as f:
    pickle.dump(emotion_encoder, f)

with open("../models/label_encoder.pkl", "wb") as f:
    pickle.dump(label_encoder, f)

with open("../models/X_features.pkl", "wb") as f:
    pickle.dump(X_features, f)

with open("../models/y_labels.pkl", "wb") as f:
    pickle.dump(y, f)
print("Feature engineering completed successfully")
