==> notebooks/06_model_training.ipynb.ipynb <==
import pickle
import numpy as np

from sklearn.model_selection import train_test_split
from sklearn.linear_model import LogisticRegression
from sklearn.naive_bayes import MultinomialNB
from sklearn.metrics import classification_report, accuracy_score


with open("../models/X_features.pkl", "rb") as f:
    X = pickle.load(f)

with open("../models/y_labels.pkl", "rb") as f:
    y = pickle.load(f)

X.shape, y.shape


X_train, X_test, y_train, y_test = train_test_split(
    X,
    y,
    test_size=0.2,
    random_state=42,
    stratify=y
)


model = LogisticRegression(
    max_iter=1000,
    n_jobs=-1
)

model.fit(X_train, y_train)


y_pred = model.predict(X_test)

accuracy = accuracy_score(y_test, y_pred)
accuracy


print(classification_report(y_test, y_pred))


with open("../models/mental_health_classifier.pkl", "wb") as f:
    pickle.dump(model, f)
print("Model training completed successfully")
